+++ protocol-worker/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+async-trait = "0.1.51"
 displaydoc = "0.2"
 futures = { version = "0.3" }
 itertools = "0.10"
